initial commit
This is a project showcasing how Cloudflare's D1 serverless SQL platform
can be used to serve event information after collecting them from the
internet. It features a frontpage that reads from D1, and a /api/luma
control panel that reads from the 50 upcoming events, with an option for
the user to scrape its data.